API

Hilbert Spaces

Qumulants.HilbertSpace — Type

HilbertSpace

Abstract type for representing Hilbert spaces.

Qumulants.ProductSpace — Type

ProductSpace <: HilbertSpace

Stores a composite HilbertSpace consisting of multiple subspaces. Generally created by computing the tensor product ⊗ of subspaces.

Qumulants.FockSpace — Type

FockSpace <: HilbertSpace

HilbertSpace defining a Fock space for bosonic operators. See also: Destroy, Create

Qumulants.NLevelSpace — Type

NLevelSpace <: HilbertSpace
NLevelSpace(name::Symbol,levels,GS=1)

Define a HilbertSpace for an object consisting of N discrete energy levels. The given levels must be an integer specifying the number of levels, or an iterable collection of levels. The argument GS specifies which state should be treated as ground state and is rewritten using population conservation during simplification. See also: Transition

Examples:

julia> ha = NLevelSpace(:a,3)
ℋ(a)

julia> ha = NLevelSpace(:a,(:g,:e))
ℋ(a)

Qumulants.:⊗ — Function

⊗(spaces::HilbertSpace...)

Create a ProductSpace consisting of multiple subspaces.

Examples

julia> hf = FockSpace(:f)
ℋ(f)

julia> ha = NLevelSpace(:a,2)
ℋ(a)

julia> h = hf⊗ha
ℋ(f) ⊗ ℋ(a)

Operators

Qumulants.AbstractOperator — Type

AbstractOperator

Abstract type representing any expression involving operators.

Qumulants.BasicOperator — Type

BasicOperator <: AbstractOperator

Abstract type representing fundamental operator types.

Qumulants.OperatorTerm — Type

OperatorTerm <: AbstractOperator

Symbolic expression tree consisting of AbstractOperator and Number arguments.

Qumulants.Destroy — Type

Destroy <: BasicOperator

Bosonic operator on a FockSpace representing the quantum harmonic oscillator annihilation operator.

Qumulants.Create — Type

Create <: BasicOperator

Bosonic operator on a FockSpace representing the quantum harmonic oscillator creation operator.

Qumulants.Transition — Type

Transition <: BasicOperator
Transition(h::NLevelSpace,name::Symbol,i,j)

Fundamental operator defining a transition from level j to level i on a NLevelSpace. The notation corresponds to Dirac notation, i.e. the above is equivalent to |i⟩⟨j|.

Examples

julia> ha = NLevelSpace(:a,(:g,:e))
ℋ(a)

julia> σ = Transition(ha,:σ,:g,:e)
σge

Heisenberg

Qumulants.heisenberg — Function

heisenberg(ops::Vector,H::AbstractOperator)
heisenberg(op::AbstractOperator,H::AbstractOperator)

Compute a set of Heisenberg equations of the operators in ops under the Hamiltonian H.

heisenberg(ops::Vector,H::AbstractOperator,J::Vector;
        Jdagger::Vector=adjoint.(J),rates=ones(length(J)))
heisenberg(op::AbstractOperator,H::AbstractOperator,J::Vector;
        Jdagger::Vector=adjoint.(J),rates=ones(length(J)))

Compute the set of equations for the operators in ops under the Hamiltonian H and with loss operators contained in J. The resulting equation is equivalent to the Quantum-Langevin equation where noise is neglected.

Arguments

*ops::Vector{<:AbstractVector}: The operators of which the equations are to be computed. *H::AbstractOperatr: The Hamiltonian describing the reversible dynamics of the system. *J::Vector{<:AbstractOperator}: A vector containing the collapse operators of the system. A term of the form $\sum_i J_i^\dagger O J_i - \frac{1}{2}\left(J_i^\dagger J_i O + OJ_i^\dagger J_i\right)$ is added to the Heisenberg equation.

Optional argumentes

*Jdagger::Vector=adjoint.(J): Vector containing the hermitian conjugates of the collapse operators. *rates=ones(length(J)): Decay rates corresponding to the collapse operators in J.

Qumulants.commutator — Function

commutator(a,b; simplify=true, kwargs...)

Computes the commutator a*b - b*a of a and b. If simplify is true, the result is simplified using the simplify_operators function. Further keyword arguments are passed to simplification.

Qumulants.acts_on — Function

acts_on(op::AbstractOperator)

Shows on which Hilbert space op acts. For BasicOperator types, this returns an Integer, whereas for a OperatorTerm it returns a Vector{Int} whose entries specify all subspaces on which the expression acts.

Differential Equations

Qumulants.AbstractEquation — Type

Abstract type for equations.

Qumulants.DifferentialEquation — Type

DifferentialEquation{LHS,RHS,H,J,R} <: AbstractEquation{LHS,RHS}
DifferentialEquation(lhs,rhs,H,J,rates)

Type defining a system of differential equations, where lhs is a vector of derivatives and rhs is a vector of expressions. In addition, it keeps track of the Hamiltonian, the collapse operators and the corresponding decay rates of the system.

Fields

*lhs: Vector of operators or averages of which the derivatives are taken. *rhs: Vector of expressions to which the derivatives are equal. *hamiltonian: Operator defining the system Hamiltonian. *jumps: Vector of operators specifying the decay processes. *rates: Decay rates corresponding to the jumps.

Symbolic Numbers

Qumulants.SymbolicNumber — Type

SymbolicNumber <: Number

Abstract type for all symbolic numbers, i.e. Parameter, Average and corresponding expression trees.

Qumulants.NumberTerm — Type

NumberTerm <: SymbolicNumber

Expression tree consisting of SymbolicNumber variables.

Qumulants.Parameter — Type

Parameter <: SymbolicNumber

A parameter represented as a symbolic. See also: parameters, @parameters

Qumulants.parameters — Function

parameters(symbols::Symbol...)
paramters(s::String)

Create symbolic parameters.

Expamples

julia> ps = parameters(:a, :b)
(a, b)

julia> parameters("a b") == ps
true

Qumulants.@parameters — Macro

@parameters(ps...)

Convenience macro to quickly define symbolic parameters.

Examples

julia> @parameters ω κ
(ω, κ)

Simplification

Qumulants.simplify_operators — Function

simplify_operators(op::AbstractOperator; rewriter=default_operator_simplifier(), kwargs...)

Simplify an operator through standard algebraic rewriting, as well as using fundamental commutation relations.

Arguments

===========

op: The operator expression to be simplified.

rewriter: The rewriter used.

kwargs: Further arguments passed to SymbolicUtils.simplify.

Qumulants.simplify_constants — Function

simplify_constants(t::NumberTerm;kwargs...)

Standard simplification for SymbolicNumber types. Converts to a SymbolicUtils expression and uses its standard simplification routines for symbolic number variables.

Qumulants.expand — Function

expand(ex; rewriter=defualt_expand_simplifier(), kwargs...)

Simple wrapper around SymbolicUtils.simplify that uses a rewriter such that expressions are expanded.

Arguments

===========

ex: The expression to be expanded.

rewriter: The used rewriter.

kwargs: Further arguments passed to SymbolicUtils.simplify.

Examples

julia> @parameters p q r
(p, q, r)

julia> ex = p*(q+r) + (q+p)*(r+q)
((p*(q+r))+((q+p)*(r+q)))

julia> expand(ex)
((p*q)+(p*r)+(q*r)+(p*r)+(q*q)+(p*q))

Qumulants.substitute — Function

substitute(arg, subs; simplify=true)

Substitute the symbolic argument, i.e. any subtype to AbstractOperator or SymbolicNumber according to the substitutions stored in a Dict. Also works on DifferentialEquation. If simplify=true, the output is simplified.

Examples

julia> @parameters p
(p,)

julia> substitute(p, Dict(p=>2))
2

Average

Qumulants.Average — Type

Average <: SymbolicNumber

Symbolic number representing the average over an operator. See also: average

Qumulants.average — Function

average(::AbstractOperator)
average(::AbstractOperator,order::Int)

Compute the average of an operator. If order is given, the cumulant_expansion up to that order is computed immediately.

average(::DifferentialEquation;multithread=false)
average(::DifferentialEquation,order::Int;multithread=false)

Compute the average of a DifferentialEquation (or a set of equations). Returns a DifferentialEquation with containing the corresponding equations for averages. If order is specified, the cumulant_expansion up to that order is computed immediately. The keyword multithread specifies whether the averaging (and cumulant_expansion) should be parallelized (defaults to false).

Qumulants.cumulant_expansion — Function

cumulant_expansion(avg, order::Int)

For an Average of an operator, expand it in terms of moments up to order neglecting their joint cumulant.

See also: https://en.wikipedia.org/wiki/Cumulant#Joint_cumulants

Examples

julia> avg = average(a*b)
⟨a*b⟩

julia> cumulant_expansion(avg,1)
(⟨a⟩*⟨b⟩)

julia> avg = average(a*b*c)
⟨a*b*c⟩

julia> cumulant_expansion(avg,2)
((⟨a*b⟩*⟨c⟩)+(⟨a*c⟩*⟨b⟩)+(⟨a⟩*⟨b*c⟩)+(-2*⟨a⟩*⟨b⟩*⟨c⟩))

Optional arguments

*simplify=true: Specify whether the result should be simplified. *kwargs...: Further keyword arguments being passed to simplify_constants

Qumulants.cumulant — Function

cumulant(x,n=get_order(x);simplify=true,kwargs...)

Compute the nth cumulant of x (either an operator or an average). The output is simplified when simplify=true. Further keyword arguments are passed on to simplification.

Examples

julia> cumulant(a)
⟨a⟩

julia> cumulant(a*b)
(⟨a*b⟩+(-1*⟨a⟩*⟨b⟩))

julia> cumulant(a*b,1)
⟨a*b⟩

julia> cumulant(a*b,3)
0

Qumulants.get_order — Function

get_order(arg)

Compute the order of a given argument. This is the order used to decide whether something should be expanded using a cumulant_expansion method.

Examples

julia> get_order(a)
1

julia> get_order(a*b)
2

julia> get_order(1)
0

Function generation

Qumulants.build_ode — Function

build_ode(rhs::Vector, vs::Vector, ps=[], usym=:u,
            psym=:p, tsym=:t; set_unknowns_zero::Bool=false, check_bounds::Bool=false)

From a set of equations contained in eqs, generate a Meta.Expr containing the code for a function which can be directly passed to OrdinaryDiffEq in order to solve it. The variable vector u corresponds to the symbols provided in vs.

Arguments

*eqs::Vector: The vector containing the right-hand side of equations. *vs::Vector: The vector containing the left-hand side of equations. *ps=[]: List of (symbolic) parameters, which are parsed into parameters used in DiffEq functions. *usym=:u: The symbol used for the variable vector. *psym=:p: The symbol used for the parameter vector. *tsym=:t: The symbol used for the time parameter.

Optional arguments

*check_bounds::Bool=false: Choose whether the resulting function should contain the @inbounds flag, which skips bounds checking for performance.

build_ode(eqs::DifferentialEquation, ps=[], usym=:u,
            psym=:p, tsym=:t; set_unknowns_zero::Bool=false, check_bounds::Bool=false)

From a set of differential equationseqs of averages, generate a Meta.Expr containing the code for a function which can be directly passed to OrdinaryDiffEq in order to solve it.

Arguments

*eqs::DifferentialEquation: The set of (average) equations. *ps=[]: List of symbolic parameters, which are parsed into parameters used in DiffEq functions. *usym=:u: The symbol used for the variable vector. *psym=:p: The symbol used for the parameter vector. *tsym=:t: The symbol used for the time parameter.

Optional arguments

*check_bounds::Bool=false: Choose whether the resulting function should contain the @inbounds flag, which skips bounds checking for performance.

Qumulants.generate_ode — Function

generate_ode(eqs::DifferentialEquation, ps=[], usym=:u,
            psym=:p, tsym=:t; set_unknowns_zero::Bool=false, check_bounds::Bool=false)

From a set of differential equations eqs of averages, generate a Function which can be directly used in OrdinaryDiffEq. Essentially, this calls Meta.eval on the output of the build_ode function.

Arguments

*eqs::DifferentialEquation: The set of (average) equations. *ps=[]: List of symbolic parameters, which are parsed into parameters used in DiffEq functions. *usym=:u: The symbol used for the variable vector. *psym=:p: The symbol used for the parameter vector. *tsym=:t: The symbol used for the time parameter.

Optional arguments

*check_bounds::Bool=false: Choose whether the resulting function should contain the @inbounds flag, which skips bounds checking for performance.

Related methods

generate_ode(eqs::Vector, vs::Vector, ps=[], usym=:u,
            psym=:p, tsym=:t; set_unknowns_zero::Bool=false, check_bounds::Bool=false)

Correlation functions

Qumulants.CorrelationFunction — Type

struct CorrelationFunction

Type representing the two-time first-order correlation function of two operators.

Qumulants.Spectrum — Type

struct Spectrum

Type representing the spectrum, i.e. the Fourier transform of a CorrelationFunction in steady state.

To actually compute the spectrum at a frequency ω, construct the type on top of a correlation function and call it with Spectrum(c)(ω,usteady,p0).

Qumulants.initial_values — Function

initial_values(c::CorrelationFunction, u_end)

Find the vector containing the correct initial values when numerical solving the time evolution for the correlation function.

When computing the correlation function of two operators in a system that has been evolved up to a time t_end, such that its state is given by u_end, this function provides the correct initial values in the right order that can be used to solve the ordinary differential equation together with the function generated by generate_ode(c).

See also: CorrelationFunction generate_ode

Utility functions

Qumulants.find_missing — Function

find_missing(rhs::Vector, vs::Vector, vs_adj=adjoint.(vs), ps=[])

For a list of expressions contained in rhs, check whether all occurring symbols are contained either in the variables given in vs. If a list of parameters ps is provided, parameters that do not occur in the list ps are also added to the list. Returns a list of missing symbols.

Qumulants.get_symbolics — Function

get_symbolics(ex)

Find all symbolic numbers occuring in ex.

Qumulants.get_operators — Function

get_operators(::AbstractOperator)

Return a list of all BasicOperator in an expression.

Qumulants.find_operators — Function

find_operators(::HilbertSpace, order; names=nothing)

Find all operators that fully define a system up to the given order.

Qumulants.complete — Function

complete(de::DifferentialEquation)

From a set of differential equation of averages, find all averages that are missing and derive the corresponding equations of motion.

Qumulants.unique_ops — Function

unique_ops(ops)

For a given list of operators, return only unique ones taking into account their adjoints.

Qumulants.fundamental_operators — Function

fundamental_operators(::HilbertSpace)

Return all fundamental operators for a given Hilbertspace. For example, a FockSpace only has one fundamental operator, Destroy.

Qumulants.transition_superscript — Function

transition_superscript(::Bool)

Specify whether the indices in a Transition operator should be printed as superscript. Default is true. If set to false, the indices corresponding to the levels are printed as subscript.